split form into separate component and wire save button
