+++ src/views/ProjectManagement/components/Users/PMUsersTab/UserProfile/index.js
@@ -1,64 +1,31 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import { reduxForm, FormSection } from 'redux-form';
-import { Tabs, Tab } from 'grommet';
+import { submit } from 'redux-form';
 
 import Modal from '../../../../../../common/Modal';
-import UserNameInput from './UserNameInput';
-import AccountTab from './AccountTab';
-import UserGroupsTab from './UserGroupsTab';
-import TasksTab from './TasksTab';
-import PreferenceTab from './PreferenceTab';
+import UserProfileForm from './UserProfileForm';
 
 import {
-    setUserFirstName,
-    setUserLastName,
-    setUserEmail,
-    setUserTitle,
+    updateUser,
 } from '../../../../../../common/actions/userActions';
-
-class UserProfileForm extends Component {
-    render() {
-        return (
-            <form>
-                <FormSection name='name'>
-                    <UserNameInput {...this.props}
-                        onFirstNameChange={this.props.onSetFirstName}
-                        onLastNameChange={this.props.onSetLastName}/>
-                </FormSection>
-                <Tabs>
-                    <Tab title={this.props.vocab.COMMON.ACCOUNT}>
-                        <FormSection name='account'>
-                            <AccountTab {...this.props}
-                                onEmailChange={this.props.onSetEmail}
-                                onTitleChange={this.props.onSetTitle}/>
-                        </FormSection>
-                    </Tab>
-                    <Tab title={this.props.vocab.PROJECT.USER_GROUPS}>
-                        <UserGroupsTab {...this.props} />
-                    </Tab>
-                    <Tab title={this.props.vocab.PROJECT.TASKS}>
-                        <TasksTab {...this.props} />
-                    </Tab>
-                    <Tab title={this.props.vocab.COMMON.PREFERENCE}>
-                        <PreferenceTab />
-                    </Tab>
-                </Tabs>
-            </form>
-        );
-    }
-}
-
-const UserProfileFormRedux = reduxForm({ form: 'user-profile' })(UserProfileForm);
-
 
 class UserProfile extends Component {
     render() {
         return (
             <Modal title={this.props.vocab.PROJECT.USER_PROFILE}
-                onCancel={this.props.onCancel}>
-                <UserProfileFormRedux {...this.props}/>
+                onCancel={this.props.onCancel}
+                onSave={this.props.onClickToSubmit}>
+                <UserProfileForm {...this.props}
+                    onSubmit={ (values) => {
+                        this.props.onSave();
+                        this.props.onUpdateUser(this.props.userId, {
+                            firstName: values.name.firstName,
+                            lastName: values.name.lastName,
+                            email: values.account.email,
+                            title: values.account.title,
+                        });
+                    }}/>
             </Modal>
         );
     }
@@ -92,11 +59,9 @@ const mapStateToProps = (state, ownProps) => {
     };
 };
 
-const mapDispatchToProps = (dispatch, ownProps) => ({
-    onSetFirstName: firstName => dispatch(setUserFirstName(ownProps.userId, firstName)),
-    onSetLastName: lastName => dispatch(setUserLastName(ownProps.userId, lastName)),
-    onSetEmail: email => dispatch(setUserEmail(ownProps.userId, email)),
-    onSetTitle: title => dispatch(setUserTitle(ownProps.userId, title)),
+const mapDispatchToProps = dispatch => ({
+    onUpdateUser: (userId, user) => dispatch(updateUser(userId, user)),
+    onClickToSubmit: () => dispatch(submit('user-profile')),
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(UserProfile);
